arelle_service/requirements.txt
boto3==1.34.149
flake8==7.1.0
flake8-noqa==1.4.0

mypy==1.11.0
pytest==8.3.2

boto3-stubs==1.34.149
typing-extensions==4.12.2
lxml-stubs==0.5.1
types-PyMySQL==1.1.0.20240524
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-simplejson==3.19.0.20240310
types-ujson==5.10.0.20240515
types-regex==2024.7.24.20240726
types-waitress==3.0.0.20240423
ixbrl-viewer
regex
arelle-release==2.30.19
holidays==0.54
pytz==2024.1
fastapi==0.112.0
python-multipart==0.0.9
uvicorn==0.30.5